+++ Controller/Adminhtml/Index/Index.php
@@ -16,9 +16,12 @@ use Magento\Backend\App\Action;
 use Magento\Backend\App\Action\Context;
 use Magento\Backend\Model\View\Result\Page;
 use Magento\Framework\App\RequestInterface;
+use Magento\Framework\App\ResponseInterface;
+use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Message\ManagerInterface;
 use Magento\Framework\View\Result\PageFactory;
 use Yireo\EmailTester2\Config\Config;
+use Yireo\EmailTester2\ViewModel\Form;
 
 /**
  * Class Index
@@ -36,31 +39,41 @@ class Index extends Action
      * @var PageFactory
      */
     private $resultPageFactory;
+
     /**
      * @var Config
      */
     private $config;
+    /**
+     * @var Form
+     */
+    private $formViewModel;
 
     /**
      * @param Context $context
      * @param PageFactory $resultPageFactory
+     * @param ManagerInterface $messageManager
+     * @param Config $config
+     * @param Form $formViewModel
      */
     public function __construct(
         Context $context,
         PageFactory $resultPageFactory,
         ManagerInterface $messageManager,
-        Config $config
+        Config $config,
+        Form $formViewModel
     ) {
         parent::__construct($context);
         $this->resultPageFactory = $resultPageFactory;
         $this->messageManager = $messageManager;
         $this->config = $config;
+        $this->formViewModel = $formViewModel;
     }
 
     /**
      * @param RequestInterface $request
      *
-     * @return \Magento\Framework\App\ResponseInterface
+     * @return ResponseInterface
      */
     public function dispatch(RequestInterface $request)
     {
@@ -77,6 +90,7 @@ class Index extends Action
      * Index action
      *
      * @return Page
+     * @throws LocalizedException
      */
     public function execute(): Page
     {
@@ -85,6 +99,18 @@ class Index extends Action
         $resultPage->setActiveMenu('Yireo_EmailTester2::index');
         $resultPage->addBreadcrumb(__('Yireo EmailTester'), __('Yireo EmailTester'));
         $resultPage->getConfig()->getTitle()->prepend(__('Yireo EmailTester'));
+
+        if ($this->formViewModel->hasCustomers() === false) {
+            $this->messageManager->addWarningMessage(__('Please add some customers to your shop first'));
+        }
+
+        if ($this->formViewModel->hasProducts() === false) {
+            $this->messageManager->addWarningMessage(__('Please add some products to your shop first'));
+        }
+
+        if ($this->formViewModel->hasOrders() === false) {
+            $this->messageManager->addWarningMessage(__('Please add some orders to your shop first'));
+        }
 
         return $resultPage;
     }