+++ Controller/Adminhtml/Index/Preview.php
@@ -13,9 +13,14 @@ namespace Yireo\EmailTester2\Controller\Adminhtml\Index;
 
 use Magento\Backend\App\Action;
 use Magento\Backend\App\Action\Context;
+use Magento\Backend\Model\View\Result\Page;
+use Magento\Backend\Model\View\Result\RedirectFactory;
+use Magento\Framework\Controller\ResultInterface;
+use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\View\Page\Config;
 use Magento\Framework\View\Result\PageFactory;
 use Yireo\EmailTester2\Block\Adminhtml\Preview as PreviewBlock;
+use Yireo\EmailTester2\ViewModel\Form;
 
 /**
  * Class Index
@@ -38,28 +43,50 @@ class Preview extends Action
     private $pageConfig;
 
     /**
+     * @var RedirectFactory
+     */
+    private $redirectFactory;
+
+    /**
+     * @var Form
+     */
+    private $formViewModel;
+
+    /**
      * @param Context $context
      * @param PageFactory $resultPageFactory
      * @param Config $pageConfig
+     * @param RedirectFactory $redirectFactory
+     * @param Form $formViewModel
      */
     public function __construct(
         Context $context,
         PageFactory $resultPageFactory,
-        Config $pageConfig
+        Config $pageConfig,
+        RedirectFactory $redirectFactory,
+        Form $formViewModel
     ) {
         parent::__construct($context);
         $this->resultPageFactory = $resultPageFactory;
         $this->pageConfig = $pageConfig;
+        $this->redirectFactory = $redirectFactory;
+        $this->formViewModel = $formViewModel;
     }
 
     /**
      * Index action
      *
-     * @return \Magento\Backend\Model\View\Result\Page
+     * @return ResultInterface
+     * @throws LocalizedException
      */
     public function execute()
     {
-        /** @var \Magento\Backend\Model\View\Result\Page $resultPage */
+        if ($this->hasValidData() === false) {
+            $this->messageManager->addWarningMessage('You have not added the required customers, products or orders yet');
+            return $this->resultRedirectFactory->create()->setPath('*/*/index');
+        }
+
+        /** @var Page $resultPage */
         $resultPage = $this->resultPageFactory->create();
 
         $this->pageConfig->getTitle()->set($this->getTemplateName());
@@ -74,6 +101,27 @@ class Preview extends Action
         $resultPage->addContent($previewBlock);
 
         return $resultPage;
+    }
+
+    /**
+     * @return bool
+     * @throws LocalizedException
+     */
+    private function hasValidData(): bool
+    {
+        if ($this->formViewModel->hasCustomers() === false) {
+            return false;
+        }
+
+        if ($this->formViewModel->hasProducts() === false) {
+            return false;
+        }
+
+        if ($this->formViewModel->hasOrders() === false) {
+            return false;
+        }
+
+        return true;
     }
 
     /**