+++ Controller/Adminhtml/Index/Send.php
@@ -14,13 +14,14 @@ namespace Yireo\EmailTester2\Controller\Adminhtml\Index;
 use Magento\Backend\App\Action;
 use Magento\Backend\App\Action\Context;
 use Magento\Catalog\Api\ProductRepositoryInterface;
-use Magento\Framework\Controller\Result\Redirect;
 use Magento\Framework\Controller\Result\RedirectFactory;
+use Magento\Framework\Controller\ResultInterface;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Framework\Message\ManagerInterface as MessageManager;
 use Magento\Store\Model\StoreManagerInterface;
 use Yireo\EmailTester2\Model\Mailer;
+use Yireo\EmailTester2\ViewModel\Form;
 
 /**
  * Class Index
@@ -53,12 +54,18 @@ class Send extends Action
     private $productRepository;
 
     /**
+     * @var Form
+     */
+    private $formViewModel;
+
+    /**
      * @param Context $context
      * @param RedirectFactory $redirectFactory
      * @param MessageManager $messageManager
      * @param Mailer $mailer
      * @param StoreManagerInterface $storeManager
      * @param ProductRepositoryInterface $productRepository
+     * @param Form $formViewModel
      */
     public function __construct(
         Context $context,
@@ -66,7 +73,8 @@ class Send extends Action
         MessageManager $messageManager,
         Mailer $mailer,
         StoreManagerInterface $storeManager,
-        ProductRepositoryInterface $productRepository
+        ProductRepositoryInterface $productRepository,
+        Form $formViewModel
     ) {
         parent::__construct($context);
         $this->redirectFactory = $redirectFactory;
@@ -74,17 +82,23 @@ class Send extends Action
         $this->messageManager = $messageManager;
         $this->storeManager = $storeManager;
         $this->productRepository = $productRepository;
+        $this->formViewModel = $formViewModel;
     }
 
     /**
      * Index action
      *
-     * @return Redirect
+     * @return ResultInterface
      * @throws LocalizedException
      * @throws NoSuchEntityException
      */
     public function execute()
     {
+        if ($this->hasValidData() === false) {
+            $this->messageManager->addWarningMessage('You have not added the required customers, products or orders yet');
+            return $this->redirectFactory->create()->setPath('*/*/index');
+        }
+
         $data = $this->getRequestData();
         $this->saveToSession($data);
 
@@ -96,6 +110,27 @@ class Send extends Action
         $redirect->setPath('*/*/index', ['form_id' => 0]);
 
         return $redirect;
+    }
+
+    /**
+     * @return bool
+     * @throws LocalizedException
+     */
+    private function hasValidData(): bool
+    {
+        if ($this->formViewModel->hasCustomers() === false) {
+            return false;
+        }
+
+        if ($this->formViewModel->hasProducts() === false) {
+            return false;
+        }
+
+        if ($this->formViewModel->hasOrders() === false) {
+            return false;
+        }
+
+        return true;
     }
 
     /**